fix SonarCloud duplicated code
